[Silver III] Title: 주유소, Time: 388 ms, Memory: 36640 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -7,18 +7,18 @@ class Main {
     public static void main(String[] args) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
-        int[] len = new int[N-1];
-        int[] price = new int[N];
-        int total = 0;
+        long[] len = new long[N-1];
+        long[] price = new long[N];
+        long total = 0;
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
         for (int i = 0; i < N-1; i++) {
-            len[i] = Integer.parseInt(st.nextToken());
+            len[i] = Long.parseLong(st.nextToken());
         }
         st = new StringTokenizer(br.readLine(), " ");
         for (int i = 0; i < N; i++) {
-            price[i] = Integer.parseInt(st.nextToken());
+            price[i] = Long.parseLong(st.nextToken());
         }
-        int min = price[0];
+        long min = price[0];
 
         int cur = 0;
         while (cur < N-1) {
